Refactor presenter lifecycles to use Loaders
This commit uses Synchronous Loaders to manage the lifecycle of
Presenters outside of the normal Activity/Fragment lifecycle.

PresenterActivity and PresenterFragment now delegation presenter
creation to PresenterLoader. One caveat of this implementation is that
each PresenterLoader requires a Presenter provider which is injected
through dagger. This means that the dependency injection in the concrete
Activity/Fragment must happen before the super.onCreate() call.